SecurityAnalyzerLLM/requirements.txt
fastapi
pydantic
langchain-core
langchain
langchain-google-genai
python-dotenv
pandas
requests
langchain_chroma